pyproject.toml
[project]
name = "hyper-jade"
version = "0.1.0"
description = "JADE system for automated code analysis and generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.1.0",
    "langchain-text-splitters>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-ollama>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
    "mlflow>=2.8.0",
    "pandas>=1.5.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "langgraph",
    "python-dotenv",
    "requests",
    "redis",
    "rq",
    "pydantic",
    "weaviate-client>=3.25.0",
    "sentence-transformers>=2.2.0",
    "nbformat>=5.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []

[project.scripts]
worker = "worker.main:main"
setup-rag = "setup_rag:main"
list-outputs = "list_outputs:main"
start-worker = "start_worker:main"
view-metrics = "view_mlflow_metrics:main"
